requirements.txt

flask
flask_cors
werkzeug
ffmpeg-python
